Perkeltas SearchSpecifications imimo duomenų algoritmas į klasę
